revising a-fun-bud tutorial before course
i corrected a few mistakes.

i found this was strange in the sense that 2 letters were not needed: For instance, by clicking on the Axial dorsal button (Figure 9c), you will get the same view as shown in Figure 9d.

also in figure 10 it is a bit odd that the letters are not in order

finally, figures 11 and 12 will have to be edited, they are quite large